Initial release pipeline
Initial version of a pipeline that will perform a release based
on a set of commits in a specific branch. The workflow will tag the
release and trigger the release procedure in other repositories, so the
assumption is that the commits for the release and the changelog is up
to date.

You can trigger this workflow either explicitly in the GitHub
interface, or using the CLI with something like:

    gh workflow run release.yaml --ref 2.1.x version=2.1.5